extract app from index.js and modify scripts to setup test env
